Move onboarding models to separate package
