+++ src/test/java/seedu/contax/model/OnboardingStepTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+import seedu.contax.model.onboarding.OnboardingStep;
 
 public class OnboardingStepTest {
