+++ src/test/java/seedu/contax/model/OnboardingStoryTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+import seedu.contax.model.onboarding.OnboardingStep;
+import seedu.contax.model.onboarding.OnboardingStory;
 
 public class OnboardingStoryTest {
